# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Centoire API",
    "version": "0.1.0",
    "description": "Centoire community platform API"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "detail"
        ]
      },
      "TagRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "PublicUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "handle": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "reputation": {
            "type": "number"
          },
          "followerCount": {
            "type": "number"
          },
          "followingCount": {
            "type": "number"
          },
          "postCount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "handle",
          "displayName",
          "avatarUrl",
          "bio",
          "reputation",
          "followerCount",
          "followingCount",
          "postCount",
          "createdAt"
        ]
      },
      "CurrentUser": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PublicUser"
          },
          {
            "type": "object",
            "properties": {
              "email": {
                "type": "string"
              },
              "role": {
                "type": "string",
                "enum": [
                  "member",
                  "admin"
                ]
              },
              "emailVerified": {
                "type": "boolean"
              },
              "onboardingCompleted": {
                "type": "boolean"
              },
              "interests": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TagRef"
                }
              }
            },
            "required": [
              "email",
              "role",
              "emailVerified",
              "onboardingCompleted",
              "interests"
            ]
          }
        ]
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "style",
              "craft",
              "business",
              "culture"
            ]
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "postCount": {
            "type": "number"
          },
          "followerCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "category",
          "description",
          "postCount",
          "followerCount"
        ]
      },
...
      "UpdateMeInput": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "handle": {
            "type": "string",
            "pattern": "^[a-z0-9_]{3,24}$"
          },
          "bio": {
            "type": "string",
            "maxLength": 160
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "SetInterestsInput": {
        "type": "object",
        "properties": {
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "maxItems": 20
          }
        },
        "required": [
          "tagIds"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/tags": {
      "get": {
        "tags": [
          "tags"
        ],
        "operationId": "listTags",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "style",
                "craft",
                "business",
                "culture"
              ]
            },
            "required": false,
            "name": "category",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "All tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Tag"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users/me": {
      "patch": {
        "tags": [
          "users"
        ],
        "operationId": "updateMe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "409": {
            "description": "Handle taken or reserved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/interests": {
      "put": {
        "tags": [
          "users"
        ],
        "operationId": "setInterests",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetInterestsInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/complete-onboarding": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "completeOnboarding",
        "responses": {
          "200": {
            "description": "Onboarding completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "422": {
            "description": "Handle or interests missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}/follow": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "followUser",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Now following"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "operationId": "unfollowUser",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "No longer following"
          }
        }
      }
    },
...
  }
}